requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
